prism-ml-auditoryhum/requirements.txt
accelerate==1.12.0
annotated-types==0.7.0
anyio==4.12.0
audioop-lts==0.2.2
audioread==3.1.0
bitsandbytes==0.49.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
filelock==3.20.1
fonttools==4.61.1
fsspec==2025.12.0
google-auth==2.45.0
google-genai==1.56.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pillow==12.0.0
platformdirs==4.5.1
pooch==1.8.2
protobuf==6.33.4
psutil==7.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
timm==1.0.22
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.6
triton==3.5.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
websockets==15.0.1
